Move audio constants to lib/constants/audio
