+++ app/containers/MessageComposer/components/RecordAudio/RecordAudio.tsx
@@ -12,13 +12,13 @@ import sharedStyles from '../../../../views/Styles';
 import { ReviewButton } from './ReviewButton';
 import { useMessageComposerApi } from '../../context';
 import { sendFileMessage } from '../../../../lib/methods';
-import { IUpload } from '../../../../definitions';
-import log from '../../../../lib/methods/helpers/log';
-import { useRoomContext } from '../../../../views/RoomView/context';
+import { RECORDING_EXTENSION, RECORDING_MODE, RECORDING_SETTINGS } from '../../../../lib/constants';
 import { useAppSelector } from '../../../../lib/hooks';
+import log from '../../../../lib/methods/helpers/log';
+import { IUpload } from '../../../../definitions';
+import { useRoomContext } from '../../../../views/RoomView/context';
 import { useCanUploadFile } from '../../hooks';
 import { Duration, IDurationRef } from './Duration';
-import { RECORDING_EXTENSION, RECORDING_MODE, RECORDING_SETTINGS } from './constants';
 import AudioPlayer from '../../../AudioPlayer';
 import { CancelButton } from './CancelButton';
 import i18n from '../../../../i18n';
